Redesign "sidecar" box
Now it's designed to be attached with 3M Dual Lock (like velcro but
stronger; when locked together about 5mm thick), and the lid is just
a piece of stiff cardboard (but could also be 3D printed).  The
sliding slot for the lid is an overhang though; maybe your printer
can cope, or maybe not.  This design is also a bit thin and not
as robust as the uradmonitor case itself.  But works for now.